Refactored socket io middleware to middleware
Also added server event "users" that sends all users
