+++ server/src/server.ts
@@ -1,61 +1,41 @@
 import app from './app';
 import {createServer, type Server as HttpServer} from 'node:http';
-import {Server, type Socket} from 'socket.io';
+import {Server} from 'socket.io';
 import logger from './utils/logger';
-import jwt from 'jsonwebtoken';
-import config from './utils/config';
-import {type UserPublic, userPublicSchema} from './validators/UserPublic';
+import {type UserPublic} from './validators/UserPublic';
+import {tokenParserMiddleware} from './utils/middleware';
+import {type SocketWithUser} from './utils/types';
 
 const server = createServer(app);
 
 export const attachSocketServerTo = (httpServer: HttpServer) => {
 	const io = new Server(httpServer);
 
-	const parseUserFromToken = (token: string) => {
-		const decodedToken: unknown = jwt.verify(token, config.jwtSecret);
-
-		const user = userPublicSchema.parse(decodedToken);
-
-		return user;
-	};
-
-	type SocketWithUser = {
-		user: UserPublic;
-	} & Socket;
-
-	let connectedUsers: UserPublic[] = [];
-
-	io.use((socket, next) => {
-		const {token} = socket.handshake.auth;
-
-		if (!token) {
-			next(new Error('Token is required'));
-			return;
-		}
-
-		if (typeof token !== 'string') {
-			next(new Error('Token in invalid format'));
-			return;
-		}
-
-		try {
-			(socket as SocketWithUser).user = parseUserFromToken(token);
-			next();
-		} catch (ex) {
-			next(new Error('Token invalid'));
-		}
-	});
+	io.use(tokenParserMiddleware);
 
 	enum SocketEvent {
 		Connection = 'connection',
 		Disconnect = 'disconnect',
 		Message = 'message',
+		Users = 'users',
 	}
 
 	type MessageContent = {
 		sender: string;
 		message: string;
 		timestamp: Date;
+	};
+
+	let connectedUsers: UserPublic[] = [];
+
+	const addUser = (user: UserPublic) => {
+		connectedUsers = connectedUsers.find(u => u.id === user.id) === undefined ? connectedUsers.concat(user) : connectedUsers;
+		io.emit(SocketEvent.Users, {connectedUsers});
+	};
+
+	const removeUser = (user: UserPublic) => {
+		connectedUsers = connectedUsers.filter(u => u.id !== user.id);
+		io.emit(SocketEvent.Users, {connectedUsers});
 	};
 
 	io.on(SocketEvent.Connection, async socket => {
@@ -71,12 +51,13 @@ export const attachSocketServerTo = (httpServer: HttpServer) => {
 		socket.on(SocketEvent.Message, (message: string) => {
 			logger.info(`message received from ${socket.id}`, message);
 			io.emit(SocketEvent.Message, {sender: user.username, message, timestamp: new Date()});
+			addUser(user);
 		});
 
 		socket.on(SocketEvent.Disconnect, () => {
 			logger.info(`${user.username} disconnected`);
 			socket.broadcast.emit(SocketEvent.Message, {sender: user.username, message: 'left the chat', timestamp: new Date()});
-			connectedUsers = connectedUsers.filter(u => u.id !== user.id);
+			removeUser(user);
 		});
 	});
